ai_experiments/New GEN RecSys/jibjob_recommender/requirements.txt
# Core data processing libraries
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2

# Deep learning frameworks
torch==2.0.1
transformers==4.30.0  # For BERT embeddings
pytorch-lightning==2.0.5  # For structured model training

# Graph neural networks
torch-geometric==2.3.1
dgl==1.1.1

# Configuration handling
pyyaml==6.0

# NLP preprocessing 
nltk==3.8.1

# Visualization (for notebooks)
matplotlib==3.7.1
seaborn==0.12.2

# Testing
pytest==7.3.1

# Utilities
tqdm==4.65.0
